feat(config): add sane defaults
Also specify `config` as a hash to make it easier to provide the config from
multiple pillar values.  Maintain backward compatibility (supplying `config` as
a string) so that the formula continues to work as expected if a `config`
string is already available.

BREAKING CHANGE: `config` can now be provided as a hash or a string; defaults
are modified and, while sane and desirable, do change the behavior of the
formula.